+++ tests/integration/mod.rs
@@ -1,6 +1,5 @@
 pub mod basic;
 pub mod edge;
 pub mod io;
-pub mod recursive;
 pub mod spec;
 pub mod value;
